clean up

+++ src/com/jcraft/jzlib/InfBlocks.java
@@ -124,7 +124,7 @@ final class InfBlocks {
 		q = read;
 
 		// compute number of bytes to copy as far as end of window
-		n = (int) ((q <= write ? write : end) - q);
+		n = ((q <= write ? write : end) - q);
 		if (n > z.avail_out)
 			n = z.avail_out;
 		if (n != 0 && r == Z_BUF_ERROR)
@@ -197,7 +197,7 @@ final class InfBlocks {
 		}
 		{
 			q = write;
-			m = (int) (q < read ? read - q - 1 : end - q);
+			m = (q < read ? read - q - 1 : end - q);
 		}
 
 		// process input based on current state
@@ -222,7 +222,7 @@ final class InfBlocks {
 					b |= (z.next_in[p++] & 0xff) << k;
 					k += 8;
 				}
-				t = (int) (b & 7);
+				t = (b & 7);
 				last = t & 1;
 
 				switch (t >>> 1) {
@@ -336,16 +336,16 @@ final class InfBlocks {
 				if (m == 0) {
 					if (q == end && read != 0) {
 						q = 0;
-						m = (int) (q < read ? read - q - 1 : end - q);
+						m = (q < read ? read - q - 1 : end - q);
 					}
 					if (m == 0) {
 						write = q;
 						r = inflate_flush(z, r);
 						q = write;
-						m = (int) (q < read ? read - q - 1 : end - q);
+						m = (q < read ? read - q - 1 : end - q);
 						if (q == end && read != 0) {
 							q = 0;
-							m = (int) (q < read ? read - q - 1 : end - q);
+							m = (q < read ? read - q - 1 : end - q);
 						}
 						if (m == 0) {
 							bitb = b;
@@ -624,7 +624,7 @@ final class InfBlocks {
 				b = bitb;
 				k = bitk;
 				q = write;
-				m = (int) (q < read ? read - q - 1 : end - q);
+				m = (q < read ? read - q - 1 : end - q);
 
 				if (last == 0) {
 					mode = TYPE;
@@ -635,7 +635,7 @@ final class InfBlocks {
 				write = q;
 				r = inflate_flush(z, r);
 				q = write;
-				m = (int) (q < read ? read - q - 1 : end - q);
+				m = (q < read ? read - q - 1 : end - q);
 				if (read != write) {
 					bitb = b;
 					bitk = k;
